# nodejs-server-server/api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "reneascca API"
#  termsOfService: "http://swagger.io/terms/"
#  contact:
#    email: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "reneascca.ca"
basePath: "/webapp/api/v1"
schemes:
- "http"
paths:
  /book:
    get:
      summary: "Get all books"
      description: "get content of all books as a collection of single search results"
      operationId: "findBooks"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Book"
  /book/{id}:
    get:
      summary: "Get book by id"
      description: "get content of book with given book id"
      operationId: "findBooksByid"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Book"
  /book/search:
    get:
      summary: "search books"
      description: "returns book contents that matches the searching criteria"
      operationId: "findBooksBySearch"
      produces:
      - "application/json"
      parameters:
      - name: "field"
        in: "query"
        required: true
        type: "string"
        enum:
        - "title"
        - "author"
        - "isbn"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Book"
#securityDefinitions:
#  petstore_auth:
#    type: "oauth2"
#    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
#    flow: "implicit"
#    scopes:
#      write:pets: "modify pets in your account"
#      read:pets: "read your pets"
#  api_key:
#    type: "apiKey"
#    name: "api_key"
#    in: "header"
definitions:
  BookContent:
    type: "object"
    properties:
      ISBN:
        type: "integer"
        format: "int64"
      Author:
        type: "string"
      Dewey:
        type: "string"
  Book:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
      cover:
        type: "string"
      content:
        $ref: "#/definitions/BookContent"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
    example:
      code: 0
      type: "type"
      message: "message"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"